Authorization

+++ server/src/routes/v1/ActivityLog/controller.ts
@@ -4,18 +4,22 @@ import { StatusCodes } from "http-status-codes";
 
 export const createActivityLog = async (req: Request, res: Response) => {
   const { action } = req.body;
+  const userID = (req as any).user?.id;
+  console.log('Activity user Id:',userID)
   if (!action) {
     return res.status(400).send("Action is required");
   }
-
+  if (!userID) {
+    return res.status(400).send("User ID is required");
+  }
   try {
-    const newActivityLog = new ActivityLog({ action, timestamp: new Date() });
+    const newActivityLog = new ActivityLog({ action, timestamp: new Date(), userID });
     await newActivityLog.save();
 
-    const count = await ActivityLog.countDocuments();
+    const count = await ActivityLog.countDocuments({userID});
     //keeping only 10 logs in the database to not make it cluttered.
     if(count > 10){
-      const logsToDelete = await ActivityLog.find().sort({ timestamp: 1 }).limit(count - 10);//checking if the count is greater than 10
+      const logsToDelete = await ActivityLog.find({userID}).sort({ timestamp: 1 }).limit(count - 10);//checking if the count is greater than 10
       const deletePromises = logsToDelete.map(log => ActivityLog.deleteOne({ _id: log._id }));//if yes then delete by log by the id.
       await Promise.all(deletePromises);
       
@@ -28,8 +32,9 @@ export const createActivityLog = async (req: Request, res: Response) => {
 };
 
 export const getActivityLogs = async (req: Request, res: Response) => {
+  const userID = (req as any).user?.id;
   try {
-    const logs = await ActivityLog.find().sort({ timestamp: -1 });
+    const logs = await ActivityLog.find({userID}).sort({ timestamp: -1 });
     res.json(logs);
   } catch (error) {
     console.error("Error fetching activity logs:", error);
